Refactoring duplication between angle and polygon graph angle-indicators
